+++ R/check_SI_computed.R
@@ -3,14 +3,14 @@
 #' @param dsn_grid list.
 #' @param output_SI character. path for output directory for SI
 #' @param SI_list character. list of SI
-#' @param siteName character. site name
+#' @param site_name character. site name
 #'
 #' @return plots list of plots
 #' @importFrom sf st_read
 #' @importFrom crsuggest suggest_crs
 #' @export
 
-check_SI_computed <- function(dsn_grid, output_SI, SI_list, siteName){
+check_SI_computed <- function(dsn_grid, output_SI, SI_list, site_name){
 
   message('Reading grid')
   aoi_grid <- sf::st_read(dsn = dsn_grid, quiet = T)
@@ -22,7 +22,7 @@ check_SI_computed <- function(dsn_grid, output_SI, SI_list, siteName){
   listSI <- ID <- list()
   for (si in SI_list){
     listSI[[si]] <- list.files(path = output_SI, pattern = si)
-    ID[[si]] <- gsub(pattern = paste0(siteName, '_'), replacement = '', x = listSI[[si]])
+    ID[[si]] <- gsub(pattern = paste0(site_name, '_'), replacement = '', x = listSI[[si]])
     ID[[si]] <- gsub(pattern = paste0('_', si, '.tiff'), replacement = '', x = ID[[si]])
   }
   ID_all <- unlist(ID)