+++ R/compute_si_from_grid.R
@@ -3,7 +3,7 @@
 #' @param rast_path character.
 #' @param mask_path character.
 #' @param plots list. list of plots (sf object)
-#' @param siteName character.
+#' @param site_name character.
 #' @param si_list character.
 #' @param output_dir character.
 #' @param overwrite boolean.
@@ -13,9 +13,12 @@
 #'
 #'
 #' @return filename_si list of file paths produced
+#' @importFrom tools file_path_sans_ext
+#' @importFrom terra vrt
+#' @importFrom progressr handlers progressor with_progress
 #' @export
 
-compute_si_from_grid <- function(rast_path, mask_path = NULL, plots, siteName, si_list,
+compute_si_from_grid <- function(rast_path, mask_path = NULL, plots, site_name, si_list,
                                  output_dir, overwrite = FALSE,
                                  spectral_bands = NULL,
                                  sensor_name = 'sentinel-2', ReflFactor = 10000){
@@ -40,9 +43,25 @@ compute_si_from_grid <- function(rast_path, mask_path = NULL, plots, siteName, s
                                             output_dir = output_dir,
                                             spectral_bands = spectral_bands,
                                             sensor_name = sensor_name,
-                                            siteName = siteName,
+                                            site_name = site_name,
                                             overwrite = overwrite,
+                                            ReflFactor = ReflFactor,
                                             p = p),
                             SIMPLIFY = F)}))
+
+  # create vrt for spectral indices
+  output_vrt <- file.path(output_dir, 'spectral_indices_vrt')
+  dir.create(path = output_vrt, showWarnings = FALSE, recursive = TRUE)
+  si_path_vrt <- list()
+  raster_name <- tools::file_path_sans_ext(basename(rast_path))
+  si_dir <- file.path(output_dir_si, 'spectral_indices')
+  for (si in si_list){
+    output_vrt_path <- file.path(output_vrt, paste0(raster_name, '_', si, '.vrt'))
+    si_path_vrt[[si]] <- lapply(output_dir_si, '[[', si)
+    v <- terra::vrt(x = unlist(si_path_vrt[[si]]),
+                    filename = output_vrt_path, overwrite = TRUE)
+  }
+
+
   return(output_dir_si)
 }