+++ R/correct_s2_stack.R
@@ -7,7 +7,7 @@
 #' @param aoi sf.
 #' @param offset_B2 boolean. value of the offset to apply to the S2 band B2
 #' @param corr_BRF boolean.
-#' @param bands2correct character. name of bands to correct from geometry
+#' @param bands_to_correct character. name of bands to correct from geometry
 #'
 #' @return list of collections per plot
 #' @importFrom terra rast writeRaster values
@@ -15,13 +15,13 @@
 #'
 correct_s2_stack <- function(s2_items, acq, raster_dir, aoi,
                             offset_B2 = F, corr_BRF = F,
-                            bands2correct = c('B8A', 'B11', 'B12')){
+                            bands_to_correct = c('B8A', 'B11', 'B12')){
   # correct for geometry of acquisition: get correcting factors
   output_dir <- dirname(raster_dir)
   if (corr_BRF)
     s2_items <- correct_geom(S2_rast = s2_items, aoi = aoi,
                              output_dir = output_dir, acq = acq,
-                             bands2correct = bands2correct)
+                             bands_to_correct = bands_to_correct)
   # harmonize bands with uniform B2 value (experimental)
   if (offset_B2){
     B2 <- s2_items$B02