+++ R/set_options_preprocS2.R
@@ -3,13 +3,13 @@
 #' @param fun character. name of the function which has optional parameters
 #' @param options list. including
 #' - cloudcover
-#' - path_S2tilinggrid
+#' - path_S2_tiling_grid
 #' - overwrite
-#' - geomAcq
+#' - geom_acq
 #' - additional_process
-#' - bands2correct
+#' - bands_to_correct
 #' - fraction_vegetation
-#' - RadiometricFilter
+#' - radiometric_filter
 #'
 #' @return options with default values when missing
 #' @export
@@ -19,20 +19,20 @@ set_options_preprocS2 <- function(fun, options = NULL){
   if (fun == 'get_s2_raster'){
     if (is.null(options$cloudcover))
       options$cloudcover <- 100
-    if (is.null(options$path_S2tilinggrid))
-      options$path_S2tilinggrid <- NULL
+    if (is.null(options$path_S2_tiling_grid))
+      options$path_S2_tiling_grid <- NULL
     if (is.null(options$overwrite))
       options$overwrite <- T
-    if (is.null(options$geomAcq))
-      options$geomAcq <- F
+    if (is.null(options$geom_acq))
+      options$geom_acq <- F
     if (is.null(options$additional_process))
       options$additional_process <- NULL
-    if (is.null(options$bands2correct))
-      options$bands2correct <- c('B8A', 'B11', 'B12')
+    if (is.null(options$bands_to_correct))
+      options$bands_to_correct <- c('B8A', 'B11', 'B12')
     if (is.null(options$fraction_vegetation))
       options$fraction_vegetation <- 5
-    if (is.null(options$RadiometricFilter))
-      options$RadiometricFilter <- NULL
+    if (is.null(options$radiometric_filter))
+      options$radiometric_filter <- NULL
   }
   return(options)
 }